Finish form structure
