+++ src/app/app.component.html
@@ -1,5 +1,5 @@
-<header></header>
-<toolbar></toolbar>
+<app-header></app-header>
+<app-toolbar></app-toolbar>
 
 <button
   mat-fab
@@ -16,16 +16,45 @@
   [mdePopoverOverlapTrigger]="true"
   [mdePopoverCloseOnClick]="false"
 >
-  <mat-card>
+  <mat-card class="card">
     <mat-tab-group>
       <mat-tab label="Food">
-        <mat-form-field>
+        <mat-form-field class="form-field">
           <mat-label>Name</mat-label>
           <input matInput />
         </mat-form-field>
-        <searchbar></searchbar>
+        <app-multiple-autocomplete
+          placeholder="Tags"
+          width="100%"
+        ></app-multiple-autocomplete>
+        <mat-form-field class="form-field">
+          <mat-label>Location</mat-label>
+          <input matInput />
+        </mat-form-field>
+        <mat-form-field class="form-field">
+          <mat-label>Date</mat-label>
+          <input matInput type="date" />
+        </mat-form-field>
+        <div>
+          <ngb-rating
+            class="rating"
+            [rate]="currentRate"
+            [max]="5"
+          ></ngb-rating>
+        </div>
+        <button mat-raised-button color="primary">Submit</button>
       </mat-tab>
-      <mat-tab label="Tag"> Content 2 </mat-tab>
+      <mat-tab label="Tag">
+        <mat-form-field class="form-field">
+          <mat-label>Name</mat-label>
+          <input matInput />
+        </mat-form-field>
+        <app-multiple-autocomplete
+          placeholder="Subtags"
+          width="100%"
+        ></app-multiple-autocomplete>
+        <button mat-raised-button color="primary">Submit</button>
+      </mat-tab>
     </mat-tab-group>
   </mat-card>
 </mde-popover>
