+++ src/app/app.component.ts
@@ -5,4 +5,7 @@ import { Component } from '@angular/core';
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss'],
 })
-export class AppComponent {}
+export class AppComponent {
+  title = 'eat-it-all';
+  currentRate = 0;
+}
