+++ src/app/app.module.ts
@@ -14,9 +14,11 @@ import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MdePopoverModule } from '@material-extended/mde';
 import { MatCardModule } from '@angular/material/card';
 import { MatTabsModule } from '@angular/material/tabs';
+import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { HeaderComponent } from './components/header/header.component';
 import { SearchbarComponent } from './components/searchbar/searchbar.component';
 import { ToolbarComponent } from './components/toolbar/toolbar.component';
+import { MultipleAutocompleteComponent } from './components/multiple-autocomplete/multiple-autocomplete.component';
 
 @NgModule({
   declarations: [
@@ -24,6 +26,7 @@ import { ToolbarComponent } from './components/toolbar/toolbar.component';
     HeaderComponent,
     SearchbarComponent,
     ToolbarComponent,
+    MultipleAutocompleteComponent,
   ],
   imports: [
     BrowserModule,
@@ -41,6 +44,7 @@ import { ToolbarComponent } from './components/toolbar/toolbar.component';
     MdePopoverModule,
     MatCardModule,
     MatTabsModule,
+    NgbModule,
   ],
   providers: [],
   bootstrap: [AppComponent],
