+++ src/app/components/multiple-autocomplete/multiple-autocomplete.component.html
@@ -1,0 +1,27 @@
+<mat-form-field [ngStyle]="{ width: width }">
+  <mat-chip-list #chipList>
+    <mat-chip
+      *ngFor="let value of selectedValues"
+      [selectable]="selectable"
+      [removable]="removable"
+      (removed)="remove(value)"
+    >
+      {{ value }}
+      <mat-icon matChipRemove *ngIf="removable">cancel</mat-icon>
+    </mat-chip>
+    <input
+      [placeholder]="placeholder"
+      #searchInput
+      [formControl]="searchCtrl"
+      [matAutocomplete]="auto"
+      [matChipInputFor]="chipList"
+      [matChipInputSeparatorKeyCodes]="separatorKeysCodes"
+      (matChipInputTokenEnd)="add($event)"
+    />
+  </mat-chip-list>
+  <mat-autocomplete #auto="matAutocomplete" (optionSelected)="selected($event)">
+    <mat-option *ngFor="let option of filteredOptions | async" [value]="option">
+      {{ option }}
+    </mat-option>
+  </mat-autocomplete>
+</mat-form-field>
