+++ src/app/components/searchbar/searchbar.component.html
@@ -1,30 +1,7 @@
-<mat-form-field class="searchbar">
-  <mat-chip-list #chipList>
-    <mat-chip
-      *ngFor="let value of selectedValues"
-      [selectable]="selectable"
-      [removable]="removable"
-      (removed)="remove(value)"
-    >
-      {{ value }}
-      <mat-icon matChipRemove *ngIf="removable">cancel</mat-icon>
-    </mat-chip>
-    <input
-      placeholder="Search"
-      #searchInput
-      [formControl]="searchCtrl"
-      [matAutocomplete]="auto"
-      [matChipInputFor]="chipList"
-      [matChipInputSeparatorKeyCodes]="separatorKeysCodes"
-      (matChipInputTokenEnd)="add($event)"
-    />
-  </mat-chip-list>
-  <mat-autocomplete #auto="matAutocomplete" (optionSelected)="selected($event)">
-    <mat-option *ngFor="let option of filteredOptions | async" [value]="option">
-      {{ option }}
-    </mat-option>
-  </mat-autocomplete>
-</mat-form-field>
+<app-multiple-autocomplete
+  width="50vw"
+  placeholder="Search"
+></app-multiple-autocomplete>
 
 <div class="search-checkbox-container">
   <mat-checkbox class="search-checkbox">Food</mat-checkbox>
