+++ src/app/components/searchbar/searchbar.component.scss
@@ -1,7 +1,3 @@
-.searchbar {
-  width: 50vw;
-}
-
 .search-checkbox {
   display: block;
 }
