+++ src/app/components/toolbar/toolbar.component.html
@@ -1,3 +1,3 @@
 <mat-toolbar class="toolbar">
-  <searchbar></searchbar>
+  <app-searchbar></app-searchbar>
 </mat-toolbar>
