+++ src/styles.scss
@@ -1,4 +1,17 @@
 /* You can add global styles to this file, and also import other style files */
 
-html, body { height: 100%; }
-body { margin: 0; font-family: Roboto, "Helvetica Neue", sans-serif; }
+html,
+body {
+  height: 100%;
+}
+body {
+  margin: 0;
+  font-family: Roboto, "Helvetica Neue", sans-serif;
+}
+
+.mat-tab-body-content {
+  overflow: hidden !important;
+}
+
+/* Importing Bootstrap SCSS file. */
+@import "~bootstrap/dist/css/bootstrap";
